Change excess topography test to C++
For consistency with the other tests
