+++ test/CMakeLists.txt
@@ -10,7 +10,7 @@ add_test(NAME random_dem COMMAND random_dem)
 set_tests_properties(random_dem PROPERTIES ENVIRONMENT_MODIFICATION
   "PATH=path_list_prepend:$<$<BOOL:${WIN32}>:$<TARGET_FILE_DIR:topotoolbox>>")
 
-add_executable(excesstopography excesstopography.c)
+add_executable(excesstopography excesstopography.cpp)
 target_link_libraries(excesstopography PRIVATE topotoolbox)
 add_test(NAME excesstopography COMMAND excesstopography)
 set_tests_properties(excesstopography PROPERTIES ENVIRONMENT_MODIFICATION
